refactor: nodes/ -> extensions/
