+++ packages/integrations/markdoc/src/extensions/shiki.ts
@@ -1,10 +1,10 @@
 // @ts-expect-error Cannot find module 'astro/runtime/server/index.js' or its corresponding type declarations.
 import { unescapeHTML } from 'astro/runtime/server/index.js';
 import type { ShikiConfig } from 'astro';
-import Markdoc, { type Schema } from '@markdoc/markdoc';
 import type * as shikiTypes from 'shiki';
+import type { AstroMarkdocConfig } from '../config.js';
+import Markdoc from '@markdoc/markdoc';
 import { getHighlighter } from 'shiki';
-import type { AstroMarkdocConfig } from '../config';
 
 // Map of old theme names to new names to preserve compatibility when we upgrade shiki
 const compatThemes: Record<string, string> = {
