--- packages/integrations/markdoc/src/nodes/index.ts
@@ -1,4 +1,0 @@
-export { setupHeadingConfig } from './heading.js';
-import { heading } from './heading.js';
-
-export const nodes = { heading };
