+++ packages/integrations/markdoc/src/runtime.ts
@@ -1,7 +1,7 @@
 import type { MarkdownHeading } from '@astrojs/markdown-remark';
 import Markdoc, { type RenderableTreeNode } from '@markdoc/markdoc';
 import type { ContentEntryModule } from 'astro';
-import { setupHeadingConfig } from './nodes/index.js';
+import { setupHeadingConfig } from './heading-ids.js';
 import type { AstroMarkdocConfig } from './config.js';
 
 /** Used to call `Markdoc.transform()` and `Markdoc.Ast` in runtime modules */
